+++ release-notes.md
@@ -2,6 +2,11 @@
 
 ## snapshot
 
+## 0.6.0
+- electrum estimate fee test
+- electrum tests
+- electrum block headers
+
 ## 0.5.2
 - electrum update